third_party/blink/web_tests/platform/generic/fast/events/keyboard-event-source-device-event-sender-expected.txt

Test that keyboard events should have a sourceCapabilities and its firesTouchEvents should be false.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

keydown
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is false
keypress
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is false
keyup
PASS event.sourceCapabilities is non-null.
PASS event.sourceCapabilities.firesTouchEvents is false
PASS successfullyParsed is true

TEST COMPLETE
